postgres_to_es/requirements.txt
certifi==2020.12.5
chardet==4.0.0
elasticsearch==7.12.0
idna==2.10
psycopg2-binary==2.8.6
python-json-logger==2.0.1
redis==3.5.3
requests==2.25.1
urllib3==1.26.4
